Created validators.py and moved validate_results_request in validators.py
