+++ codespeed/validators.py
@@ -1,0 +1,35 @@
+from django.core.exceptions import ValidationError
+
+
+def validate_results_request(data):
+    """
+    Validates that a result request dictionary has all needed parameters
+    and their type is correct.
+
+    Throws ValidationError on error
+    """
+    mandatory_data = [
+        'env',
+        'proj',
+        'branch',
+        'exe',
+        'ben',
+    ]
+
+    for key in mandatory_data:
+        if key not in data:
+            raise ValidationError('Key "' + key +
+                                  '" missing from GET request!')
+        elif data[key] == '':
+            raise ValidationError('Value for key "' + key +
+                                  '" empty in GET request!')
+
+    # Check that 'revs' is the correct format (if it exists)
+    if 'revs' in data:
+        try:
+            rev_value = int(data['revs'])
+        except ValueError:
+            raise ValidationError('Value for key "revs" is not an integer!')
+        if rev_value <= 0:
+            raise ValidationError('Value for key "revs" should be a'
+                                  ' strictly positive integer!', True)
